Spell Check
Spell check now detects if the selected word was found within the trie.
Currently, this allows partials of the target word to be returned as a
valid hit. If time permits, adding a hash table to verify validation
accuracy would improve the quality of this program.

Need to add file I/O to read in the dictionary.